Fixed post comment vote count flicker
